modified:   source/fixes/AladdinInNasirasRevengeWidescreenFix.cpp 	modified:   source/fixes/AlcatrazPrisonEscapeFOVFix.cpp 	modified:   source/fixes/AmericanMcGeesAliceFOVFix.cpp 	modified:   source/fixes/Americas10MostWantedFOVFix.cpp 	modified:   source/fixes/AnneMcCaffreysFreedomFirstResistanceFOVFix.cpp 	modified:   source/fixes/AntzExtremeRacingWidescreenFix.cpp 	modified:   source/fixes/ArthursQuestBattlefortheKingdomFOVFix.cpp 	modified:   source/fixes/AtlantisTheLostEmpireTrialbyFireFOVFix.cpp 	modified:   source/fixes/BarbieAsTheIslandPrincessFOVFix.cpp 	modified:   source/fixes/BarbieExplorerWidescreenFix.cpp 	modified:   source/fixes/BarbieHorseAdventuresMysteryRideWidescreenFix.cpp 	modified:   source/fixes/BarbieHorseAdventuresRidingCampWidescreenFix.cpp 	modified:   source/fixes/BarbieInThe12DancingPrincessesWidescreenFix.cpp 	modified:   source/fixes/BarbieSparklingIceShowWidescreenFix.cpp 	modified:   source/fixes/BarbieandtheMagicofPegasusWidescreenFix.cpp 	modified:   source/fixes/BatmanVengeanceWidescreenFix.cpp 	modified:   source/fixes/BattleIsleTheAndosiaWarFOVFix.cpp 	modified:   source/fixes/BeetleCrazyCupFOVFix.cpp 	modified:   source/fixes/BeyondNormandyAssignmentBerlinWidescreenFix.cpp 	modified:   source/fixes/BigMuthaTruckers2WidescreenFix.cpp 	modified:   source/fixes/BigMuthaTruckersFOVFix.cpp 	modified:   source/fixes/BillyBladeTempleofTimeFOVFix.cpp 	modified:   source/fixes/CTUMarineSharpshooterFOVFix.cpp 	modified:   source/fixes/Cabelas4x4OffroadAdventure3FOVFix.cpp 	modified:   source/fixes/Carmageddon2WidescreenFix.cpp 	modified:   source/fixes/ChameleonWidescreenFix.cpp 	modified:   source/fixes/CircusEmpireFOVFix.cpp 	modified:   source/fixes/CodenameSilverWidescreenFix.cpp 	modified:   source/fixes/CrisisTeamAmbulanceDriverFOVFix.cpp 	modified:   source/fixes/Croc2FOVFix.cpp 	modified:   source/fixes/CrusadersofMightandMagicFOVFix.cpp 	modified:   source/fixes/DieHardNakatomiPlazaFOVFix.cpp 	modified:   source/fixes/DrBrainActionReactionFOVFix.cpp 	modified:   source/fixes/DromeRacersFOVFix.cpp 	modified:   source/fixes/EliteForcesNavySEALsFOVFix.cpp 	modified:   source/fixes/EliteForcesNavySEALsWeaponsofMassDestructionFOVFix.cpp 	modified:   source/fixes/EliteForcesWW2IwoJimaFOVFix.cpp 	modified:   source/fixes/EliteForcesWW2NormandyFOVFix.cpp 	modified:   source/fixes/FroggerBeyondFOVFix.cpp 	modified:   source/fixes/FroggerTheGreatQuestWidescreenFix.cpp 	modified:   source/fixes/GlobalOperationsFOVFix.cpp 	modified:   source/fixes/HarleyDavidsonRacetotheRallyFOVFix.cpp 	modified:   source/fixes/HarryPotterQuidditchWorldCupFOVFix.cpp 	modified:   source/fixes/HugoBukkazoomWidescreenFix.cpp 	modified:   source/fixes/JackedWidescreenFix.cpp 	modified:   source/fixes/KISSPsychoCircusTheNightmareChildFOVFix.cpp 	modified:   source/fixes/KaoTheKangarooHUDandFOVFix.cpp 	modified:   source/fixes/LOTRFellowshipoftheRingWidescreenFix.cpp 	modified:   source/fixes/LarryRaglands4x4ChallengeFOVFix.cpp 	modified:   source/fixes/LegendsofMightandMagicFOVFix.cpp 	modified:   source/fixes/LegoRacers2WidescreenFix.cpp 	modified:   source/fixes/MadeManConfessionsOfTheFamilyBloodWidescreenFix.cpp 	modified:   source/fixes/MagicTheGatheringBattlegroundsFOVFix.cpp 	modified:   source/fixes/MechWarrior3FOVFix.cpp 	modified:   source/fixes/MiamiViceFOVChanger.cpp 	modified:   source/fixes/MightandMagic9FOVFix.cpp 	modified:   source/fixes/MistmareFOVFix.cpp 	modified:   source/fixes/NinaAgentChroniclesFOVFix.cpp 	modified:   source/fixes/OperationSandstormFOVChanger.cpp 	modified:   source/fixes/PacmanAdventuresInTimeFOVFix.cpp 	modified:   source/fixes/ProjectIGI1WidescreenFix.cpp 	modified:   source/fixes/RequiemAvengingAngelFOVFix.cpp 	modified:   source/fixes/RevolutionFOVFix.cpp 	modified:   source/fixes/RuneGoldCutscenesFOVFix.cpp 	modified:   source/fixes/SASAntiTerrorForceFOVChanger.cpp 	modified:   source/fixes/SOEOperationAvalancheFOVChanger.cpp 	modified:   source/fixes/SanityAikensArtifactFOVFix.cpp 	modified:   source/fixes/SearchandRescueCoastalHeroesFOVFix.cpp 	modified:   source/fixes/SearchandRescueVietnamMedEvacFOVFix.cpp 	modified:   source/fixes/SecretAgentBarbieWidescreenFix.cpp 	modified:   source/fixes/SecretServiceInHarmsWayFOVFix.cpp 	modified:   source/fixes/ShadowForceRazorUnitFOVFix.cpp 	modified:   source/fixes/Spiderman2TheGameFOVFix.cpp 	modified:   source/fixes/SplatMagazineRenegadePaintballFOVFix.cpp 	modified:   source/fixes/Starsky&HutchFOVFix.cpp 	modified:   source/fixes/StateofEmergencyWidescreenFix.cpp 	modified:   source/fixes/Swat3FOVChanger.cpp 	modified:   source/fixes/TNNOutdoorsProHunter2FOVFix.cpp 	modified:   source/fixes/TheChickenatorFOVFix.cpp 	modified:   source/fixes/TheHistoryChannelCivilWarANationDividedFOVFix.cpp 	modified:   source/fixes/TheMummyWidescreenFix.cpp 	modified:   source/fixes/TheSumofAllFearsFOVFix.cpp 	modified:   source/fixes/USMostWantedNowhereToHideFOVFix.cpp 	modified:   source/fixes/VietnamBlackOpsFOVFix.cpp 	modified:   source/fixes/WantedGunsWidescreenFix.cpp 	modified:   source/fixes/WesternOutlawWantedDeadorAliveFOVFix.cpp 	modified:   source/fixes/WingsOfWarFOVFix.cpp

+++ source/fixes/BarbieHorseAdventuresRidingCampWidescreenFix.cpp
@@ -23,7 +23,7 @@ string input, descriptor;
 fstream file;
 int choice1, choice2, tempChoice;
 bool fileNotFound, validKeyPressed;
-float newCameraFOV, newAspectRatio, newAspectRatioValue;
+double newCameraFOV, newAspectRatio, newAspectRatioValue;
 char ch;
 
 // Function to handle user input in choices
@@ -62,7 +62,7 @@ void HandleChoiceInput(int &choice)
     }
 }
 
-void HandleFOVInput(float &customFOV)
+void HandleFOVInput(double &customFOV)
 {
     do
     {
@@ -72,7 +72,7 @@ void HandleFOVInput(float &customFOV)
         // Replaces all commas with dots
         replace(input.begin(), input.end(), ',', '.');
 
-        // Parses the string to a float
+        // Parses the string to a double
         customFOV = stof(input);
 
         if (cin.fail())
@@ -146,9 +146,9 @@ void OpenFile(fstream &file, const string &filename)
     }
 }
 
-float NewAspectRatioCalculation(uint32_t &newWidthValue, uint32_t &newHeightValue)
+double NewAspectRatioCalculation(uint32_t &newWidthValue, uint32_t &newHeightValue)
 {
-    newAspectRatioValue = static_cast<float>(newWidthValue) / static_cast<float>(newHeightValue);
+    newAspectRatioValue = static_cast<double>(newWidthValue) / static_cast<double>(newHeightValue);
     return newAspectRatioValue;
 }
 
@@ -261,7 +261,7 @@ int main()
             break;
 
         case 2:
-            newCameraFOV = 1.169370532f;
+            newCameraFOV = 1.169370532;
             break;
         }
